requirements.txt
backoff==2.2.1
chromadb==0.4.22
dspy_ai==2.1.1
langchain==0.1.4
llama_index==0.9.38
openai==1.10.0
pandas==2.2.0
torch==2.1.2
tqdm==4.65.0
transformers==4.36.2
